// components/schemes_configuration_ui/src/Containers/NSIMeasuringInstrumentsJournal/types.ts
export const enum InstrumentSearchInputName {
	EQUIPMENT_SHORT_NAME = 'nsi-measuring-instrument-equipment-short-name',
	MANUFACTURER_TYPE_NAME = 'nsi-measuring-instrument-manufacturer-type-name',
	LOCATION = 'nsi-measuring-instrument-location',
	FACTORY_NUMBER = 'nsi-measuring-instrument-factory-number',
}